Add project structure POD
