+++ internal/domain/product/products_test.go
@@ -1,0 +1,1 @@
+package product
