+++ internal/domain/product/service.go
@@ -1,0 +1,1 @@
+package product
